+++ jkube-kit/config/service/src/main/java/org/eclipse/jkube/kit/config/service/PatchService.java
@@ -27,8 +27,8 @@ import io.fabric8.kubernetes.api.model.Service;
 import io.fabric8.kubernetes.api.model.batch.DoneableJob;
 import io.fabric8.kubernetes.api.model.batch.Job;
 import io.fabric8.kubernetes.client.KubernetesClient;
-import io.fabric8.kubernetes.api.model.apiextensions.CustomResourceDefinition;
-import io.fabric8.kubernetes.api.model.apiextensions.DoneableCustomResourceDefinition;
+import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.CustomResourceDefinition;
+import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.DoneableCustomResourceDefinition;
 import io.fabric8.openshift.api.model.BuildConfig;
 import io.fabric8.openshift.api.model.DoneableBuildConfig;
 import io.fabric8.openshift.api.model.DoneableImageStream;
@@ -218,7 +218,7 @@ public class PatchService {
             }
 
             DoneableCustomResourceDefinition entity =
-                    client.customResourceDefinitions()
+                    client.apiextensions().v1beta1().customResourceDefinitions()
                             .withName(oldObj.getMetadata().getName())
                             .edit();